requirements.txt
pip
numpy
opencv
sys
streamlit
Pillow